extend API response items for provenance info on all elements
