+++ src/main/java/org/sbml4j/model/api/provenance/ProvenanceInfoItem.java
@@ -35,7 +35,7 @@ import io.swagger.annotations.ApiModelProperty;
 Generated(value = "org.openapitools.codegen.languages.SpringCodegen",
           date = "2022-04-02T23:09:46.842042+02:00[Europe/Berlin]")
 @JsonInclude(Include.NON_NULL)
-public class ProvenanceInfoItem {
+public class ProvenanceInfoItem extends ProvenanceApiItem {
   @JsonProperty("type") private String type;
 
   @JsonProperty("contents")
